Move commands to the brainvisa.maker.commands module
