+++ python/brainvisa/maker/installer.py
@@ -9,7 +9,6 @@ anymore since BrainVISA 5.0.
 from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 
-import datetime
 import distutils
 import glob
 from optparse import OptionParser
@@ -23,20 +22,15 @@ import sys
 import tempfile
 import time
 import six
-from socket import gethostname
 
 from brainvisa.maker import build
-import brainvisa.maker.commands
 from brainvisa.maker.commands import IGNORED_STEP
 import brainvisa.maker.configuration
 from brainvisa.maker.environment import environmentPathVariablesSubstitution
 from brainvisa.maker.environment import normalize_path
 from brainvisa.maker.subprocess import subprocess32
 from brainvisa.maker.subprocess import system_output_on_error
-
-
-_installer_datetime = None
-_installer_variables = None
+from brainvisa.maker.utils import global_installer_datetime
 
 
 def get_matching_dirs(directories, pattern):
@@ -48,410 +42,6 @@ def get_matching_dirs(directories, pattern):
 
     return dirs
 
-
-def get_standard_arch(arch = platform.architecture()[0]):
-    return 64 if arch in ['64', '64bit', 'x86_64'] else 32
-
-
-def get_host_system_name():
-    systems = {'darwin' : 'osx',
-               'windows': 'win'}
-    system = platform.system()
-    osname = system.lower()
-    osname = systems.get(osname, osname)
-
-    if osname in ('linux', 'win'):
-        # Append architecture
-        arch = platform.architecture()[0]
-        osname += str(get_standard_arch(arch))
-
-    return osname
-
-
-def get_host_libc_version():
-    # determine libc version - using ctypes and calling C
-    # gnu_get_libc_version() function
-    # Note: plaform.libc_ver() is completely bogus.
-    import ctypes
-    libc = ctypes.cdll.LoadLibrary("libc.so.6")
-    gnu_get_libc_version = libc.gnu_get_libc_version
-    gnu_get_libc_version.restype = ctypes.c_char_p
-
-    ver = gnu_get_libc_version()
-    if sys.version_info[0] >= 3:
-        # in python3, ver is a bytes, not a str
-        ver = ver.decode()
-    return ver.split('.')
-
-
-def get_pack_host_system_name():
-    '''
-        Get system name to use for packaging.
-        Because linux compatibility for packs depends on libc version, the
-        libc version is integrated to the system name.
-        On Mac, the version of OSX is also appended.
-    '''
-    pack_system = get_host_system_name()
-    if pack_system.startswith('linux'):
-        libc_version = get_host_libc_version()
-        pack_system += '-glibc-'+ '.'.join(libc_version[:2])
-    elif pack_system.startswith('osx'):
-        pack_system += '-' + '.'.join(platform.mac_ver()[0].split('.')[:2])
-
-    return pack_system
-
-
-def installer_parse_date(value):
-    return datetime.datetime.strptime(value, '%Y_%m_%d').timetuple()[:3]
-
-
-def installer_parse_time(value):
-    return datetime.datetime.strptime(value, '%H:%M:%S').timetuple()[3:6]
-
-
-def installer_format_date(date):
-    return '%04d_%02d_%02d' % date
-
-
-def installer_format_time(time):
-    return '%02d:%02d:%02d' % time
-
-
-def global_installer_datetime():
-    global _installer_datetime
-    if _installer_datetime is not None:
-        return _installer_datetime
-
-    plt = time.localtime()
-    p_date = installer_format_date(plt[:3])
-    p_time = installer_format_time(plt[3:6])
-
-    _installer_datetime = {'date': p_date, 'time': p_time}
-
-    return _installer_datetime
-
-
-def global_installer_variables():
-    global _installer_variables
-    if _installer_variables is not None:
-        return _installer_variables
-
-    pack_host_system = get_pack_host_system_name()
-
-    _installer_variables = {'os': pack_host_system,
-                            'hostname': gethostname().split('.')[0]}
-    return _installer_variables
-
-
-class PackCommand(brainvisa.maker.commands.StepCommand):
-
-    def __init__(self, argv, configuration):
-        usage = '''%prog [global options] pack [options]
-
-    Make installer package for the selected build directory.'''
-        parser = OptionParser(usage=usage)
-        parser.add_option('--only-if-default', dest='in_config',
-                          action='store_true',
-                          default=False,
-                          help='only perform this step if it is a default '
-                          'step, or specified in the "default_steps" option '
-                          'of bv_maker.cfg config file. Default steps are '
-                          'normally "sources" for source sections, '
-                          '"configure build" for build sections.')
-        (options, args) = parser.parse_args(argv)
-        if args:
-            raise ValueError('Invalid option: %s' % args[0])
-
-        super(PackCommand, self).__init__(argv, configuration)
-        self.python_vars = dict(global_installer_datetime())
-        self.options = options
-        self.args = args
-
-    def __call__(self):
-        # First, we need to order packages to manage dependencies between
-        # them (especially for offline installer that refer to data package,
-        # because in this case, it is necessary to update data package
-        # repository before software package has been packaged)
-        def __getPackageDirectoriesByDepth():
-            def __getDepth(package_dir):
-                data_dir = package_dir.get_data_dir()
-                if not data_dir:
-                    return 0
-                else:
-                    return __getDepth(data_dir) + 1
-
-            dirs=[]
-            for o in self.configuration.packageDirectories.values():
-                dirs.append((__getDepth(o), o))
-
-            #print([(o, d.directory) for o, d in sorted(dirs)])
-            return [d for o, d in sorted(dirs, key=lambda x: x[0])]
-
-        self.process('pack', __getPackageDirectoriesByDepth(),
-                     'package', self.options, self.args)
-
-
-class InstallPackCommand(brainvisa.maker.commands.StepCommand):
-
-    def __init__(self, argv, configuration):
-        usage = '''%prog [global options] install_pack [options]
-
-    Install a binary package for the selected build directory.'''
-        parser = OptionParser(usage=usage)
-        parser.add_option('--only-if-default', dest='in_config',
-                          action='store_true',
-                          default=False,
-                          help='only perform this step if it is a default '
-                          'step, or specified in the "default_steps" option '
-                          'of bv_maker.cfg config file. Default steps are '
-                          'normally "sources" for source sections, '
-                          '"configure build" for build sections.')
-        parser.add_option('--package-date', dest='package_date',
-                          default=None,
-                          help='sets the date of the pack to install. '
-                          'This is only useful if a %(date)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-time', dest='package_time',
-                          default=None,
-                          help='sets the time of the pack to install. '
-                          'This is only useful if a %(time)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-version', dest='package_version',
-                          default=None,
-                          help='sets the version of the pack to install. '
-                          'This is only useful if a %(version)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--prefix', dest='prefix',
-                          default=None,
-                          help='sets the prefix directory to install the pack.')
-        parser.add_option('--local', dest='local',
-                          action='store_true',
-                          default=False,
-                          help='True if the installation must be done ' \
-                               'locally. Default is False.')
-        parser.add_option('--offline', dest='offline',
-                          action='store_true',
-                          default=False,
-                          help='True if the installation must be done using ' \
-                               'offline installer. Default is False.')
-        parser.add_option('--debug', dest='debug',
-                          action='store_true',
-                          default=False,
-                          help='True if the installation must be done in debug ' \
-                               'mode (i.e. generated files must not be deleted). ' \
-                               'Default is False.')
-        (options, args) = parser.parse_args(argv)
-        if args:
-            raise ValueError('Invalid option: %s' % args[0])
-
-        super(InstallPackCommand, self).__init__(argv, configuration)
-
-        date = installer_format_date(installer_parse_date(options.package_date)) \
-               if options.package_date else global_installer_datetime()['date']
-        time = installer_format_time(installer_parse_time(options.package_time)) \
-               if options.package_time else global_installer_datetime()['time']
-        self.python_vars = {'date': date,
-                            'time': time}
-        if options.package_version:
-            self.python_vars.update({'version': options.package_version})
-
-        self.options = options
-        self.args = args
-
-    def __call__(self):
-        self.process('install_pack', list(self.configuration.packageDirectories.values()),
-                     'install_package', self.options, self.args)
-
-
-class TestPackCommand(brainvisa.maker.commands.StepCommand):
-
-    def __init__(self, argv, configuration):
-        usage = '''%prog [global options] test_pack [options]
-
-    Test in installed package for the selected build directory.'''
-        parser = OptionParser(usage=usage)
-        parser.add_option('--only-if-default', dest='in_config',
-                          action='store_true',
-                          default=False,
-                          help='only perform this step if it is a default '
-                          'step, or specified in the "default_steps" option '
-                          'of bv_maker.cfg config file. Default steps are '
-                          'normally "sources" for source sections, '
-                          '"configure build" for build sections.')
-        parser.add_option('-t', '--ctest_options',
-                          default=None,
-                          help='options passed to ctest (ex: "-VV -R carto*"). '
-                          'Same as the configuration option ctest_options but '
-                          'specified at runtime. The commandline option here '
-                          'overrides the bv_maker.cfg options.')
-        parser.add_option('--package-date', dest='package_date',
-                          default=None,
-                          help='sets the date of the pack to install. '
-                          'This is only useful if a %(date)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-time', dest='package_time',
-                          default=None,
-                          help='sets the time of the pack to install. '
-                          'This is only useful if a %(time)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-version', dest='package_version',
-                          default=None,
-                          help='sets the version of the pack to install. '
-                          'This is only useful if a %(version)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        (options, args) = parser.parse_args(argv)
-        if args:
-            raise ValueError('Invalid option: %s' % args[0])
-
-        super(TestPackCommand, self).__init__(argv, configuration)
-
-        date = installer_format_date(installer_parse_date(options.package_date)) \
-               if options.package_date else global_installer_datetime()['date']
-        time = installer_format_time(installer_parse_time(options.package_time)) \
-               if options.package_time else global_installer_datetime()['time']
-        self.python_vars = {'date': date,
-                            'time': time}
-        if options.package_version:
-            self.python_vars.update({'version': options.package_version})
-
-        self.options = options
-        self.args = args
-
-    def __call__(self):
-        self.process('test_pack', list(self.configuration.packageDirectories.values()),
-                     'test_package', self.options, self.args)
-
-
-class TestrefPackCommand(brainvisa.maker.commands.StepCommand):
-
-    def __init__(self, argv, configuration):
-        usage = '''%prog [global options] testref_pack [options]
-
-    Create test reference files in installed package for the selected build directory.'''
-        parser = OptionParser(usage=usage)
-        parser.add_option('--only-if-default', dest='in_config',
-                          action='store_true',
-                          default=False,
-                          help='only perform this step if it is a default '
-                          'step, or specified in the "default_steps" option '
-                          'of bv_maker.cfg config file. Default steps are '
-                          'normally "sources" for source sections, '
-                          '"configure build" for build sections.')
-        parser.add_option('-m', '--make_options',
-                          default=None,
-                          help='options passed to make (ex: "-j8") during test '
-                          'reference generation. '
-                          'Same as the configuration option make_options but '
-                          'specified at runtime. The commandline option here '
-                          'overrides the bv_maker.cfg options.')
-        parser.add_option('--package-date', dest='package_date',
-                          default=None,
-                          help='sets the date of the pack to install. '
-                          'This is only useful if a %(date)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-time', dest='package_time',
-                          default=None,
-                          help='sets the time of the pack to install. '
-                          'This is only useful if a %(time)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-version', dest='package_version',
-                          default=None,
-                          help='sets the version of the pack to install. '
-                          'This is only useful if a %(version)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        (options, args) = parser.parse_args(argv)
-        if args:
-            raise ValueError('Invalid option: %s' % args[0])
-
-        super(TestrefPackCommand, self).__init__(argv, configuration)
-
-        date = installer_format_date(installer_parse_date(options.package_date)) \
-               if options.package_date else global_installer_datetime()['date']
-        time = installer_format_time(installer_parse_time(options.package_time)) \
-               if options.package_time else global_installer_datetime()['time']
-        self.python_vars = {'date': date,
-                            'time': time}
-        if options.package_version:
-            self.python_vars.update({'version': options.package_version})
-
-        self.options = options
-        self.args = args
-
-    def __call__(self):
-        self.process('testref_pack', list(self.configuration.packageDirectories.values()),
-                     'testref_package', self.options, self.args)
-
-
-class PublishPackCommand(brainvisa.maker.commands.StepCommand):
-
-    def __init__(self, argv, configuration):
-        usage = '''%prog [global options] publish [options]
-
-    Run command to publish package for the selected publication directory.'''
-        parser = OptionParser(usage=usage)
-        parser.add_option('--only-if-default', dest='in_config',
-                          action='store_true',
-                          default=False,
-                          help='only perform this step if it is a default '
-                          'step, or specified in the "default_steps" option '
-                          'of bv_maker.cfg config file. Default steps are '
-                          'normally "sources" for source sections, '
-                          '"configure build" for build sections.')
-        parser.add_option('-m', '--make_options',
-                          default=None,
-                          help='options passed to make (ex: "-j8") during test '
-                          'reference generation. '
-                          'Same as the configuration option make_options but '
-                          'specified at runtime. The commandline option here '
-                          'overrides the bv_maker.cfg options.')
-        parser.add_option('--package-date', dest='package_date',
-                          default=None,
-                          help='sets the date of the pack to install. '
-                          'This is only useful if a %(date)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-time', dest='package_time',
-                          default=None,
-                          help='sets the time of the pack to install. '
-                          'This is only useful if a %(time)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        parser.add_option('--package-version', dest='package_version',
-                          default=None,
-                          help='sets the version of the pack to install. '
-                          'This is only useful if a %(version)s pattern '
-                          'has been used in the package directory sections '
-                          'of bv_maker.cfg.')
-        (options, args) = parser.parse_args(argv)
-        if args:
-            raise ValueError('Invalid option: %s' % args[0])
-
-        super(PublishPackCommand, self).__init__(argv, configuration)
-
-        date = installer_format_date(installer_parse_date(options.package_date)) \
-               if options.package_date else global_installer_datetime()['date']
-        time = installer_format_time(installer_parse_time(options.package_time)) \
-               if options.package_time else global_installer_datetime()['time']
-        self.python_vars = {'date': date,
-                            'time': time}
-        if options.package_version:
-            self.python_vars.update({'version': options.package_version})
-
-        self.options = options
-        self.args = args
-
-    def __call__(self):
-        self.process('publish_pack', list(self.configuration.publicationDirectories.values()),
-                     'publish_package', self.options, self.args)
 
 
 class PackageDirectory(build.ComponentsConfigParser,
